agents/pyproject.toml
[project]
name = "adam"
version = "0.1.0"
description = "Adam assistant, a personal assistant for your daily tasks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.16",
    "dotenv>=0.9.9",
    "google-adk>=0.1.0",
    "google-genai>=1.10.0",
    "gunicorn>=23.0.0",
    "kaleido>=1.0.0",
    "litellm>=1.73.2",
    "matplotlib>=3.10.3",
    "mcp[cli]>=1.6.0",
    "plotly>=6.2.0",
    "requests>=2.32.3",
    "starlette>=0.46.2",
    "uvicorn>=0.35.0",
]
